+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 
 ### Added
 
-* `Faker.Bank` [@tobyhinloopen][]
+* `Faker.Code.iban` [@tobyhinloopen][]
 * `Faker.Beer` [@orieken][]
 * `Faker.Date` [@tobyhinloopen][]
 * `Faker.Nato` [@petehamilton][]